ghost_protocol/requirements.txt
# Ghost Protocol - Single Packet Authorization
# Dependencies for client and server components

# Cryptography
pynacl>=1.5.0  # NaCl/Sodium SecretBox for authenticated encryption

# Packet manipulation (server only)
scapy>=2.5.0  # Packet sniffing and crafting
